# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WalletLens API",
    "version": "1.0.0",
    "description": "Paid x402 API for EVM wallet reports, normalized portfolio snapshots, and enriched transaction history."
  },
  "servers": [
    {
      "url": "https://walletlens.wallyweb.com"
    }
  ],
  "paths": {
...
    "/ask": {
      "get": {
        "summary": "Convert natural-language wallet intent into a paid WalletLens URL",
        "description": "Free helper for agent workflows. Use when a user asks to analyze a wallet, summarize a Base wallet, inspect token holdings, find USDC transfers, or get transaction history. The response detects an EVM address from q, recommends /wallet-report, /portfolio, or /tx-history, and returns quoteUrl plus paidUrl.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "examples": [
                "analyze wallet 0x52E29e0d2Aa49bfBfC548C0A9F2196F4aa51f3ea on base",
                "show recent USDC transfers for 0x52E29e0d2Aa49bfBfC548C0A9F2196F4aa51f3ea",
                "get token balances for 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045 on ethereum"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Intent helper response with detected address, recommended endpoint, quoteUrl, and paidUrl",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/analyze": {
      "get": {
        "summary": "Build a paid WalletLens URL from structured wallet params",
        "description": "Free helper for agents that already have an EVM address. Returns quoteUrl, paidUrl, price, network, and x402 next steps.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$",
              "examples": ["0x52E29e0d2Aa49bfBfC548C0A9F2196F4aa51f3ea"]
            },
            "description": "EVM wallet address."
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "base",
              "examples": ["base", "base,ethereum"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Intent helper response with quoteUrl, paidUrl, price, network, and x402 next steps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/examples": {
      "get": {
        "summary": "Examples page or machine-readable examples",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["json"]
            },
            "description": "Use format=json for agent-readable examples."
          }
        ],
        "responses": {
          "200": {
            "description": "Examples HTML or JSON",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/quote": {
      "get": {
        "summary": "Free quote and call guidance",
        "description": "Validates an intended wallet analysis request and returns price, payment network, required parameters, and copy-paste paid endpoint examples.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$"
            }
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "base"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/portfolio": {
      "get": {
        "summary": "Get a normalized wallet portfolio snapshot",
        "description": "Requires x402 payment. Use this endpoint when the user asks for EVM token balances, wallet holdings, stablecoin totals, top holdings, or a normalized portfolio snapshot. Unpaid requests with a valid address return HTTP 402 with a payment-required header. If you do not have an address yet, call /ask or /analyze first.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$",
              "examples": ["0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"]
            },
            "description": "EVM wallet address."
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "base,ethereum",
              "examples": ["base", "base,ethereum", "base,ethereum,arbitrum,optimism,polygon"]
            },
            "description": "Comma-separated supported chain slugs."
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio snapshot after successful x402 payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Payment required",
            "headers": {
              "payment-required": {
                "description": "Base64-encoded x402 payment requirements.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Server or provider error"
          }
        }
      }
    },
    "/tx-history": {
      "get": {
        "summary": "Get enriched wallet transaction history",
        "description": "Requires x402 payment. Use this endpoint when the user asks for recent wallet activity, transaction history, USDC transfers, decoded transfer rows, counterparties, or Base transfer history. Returns normalized multi-chain transfer history with direction, categorization, decoded summaries, labels, and basic insights. If you do not have an address yet, call /ask or /analyze first.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$",
              "examples": ["0x52E29e0d2Aa49bfBfC548C0A9F2196F4aa51f3ea"]
            },
            "description": "EVM wallet address."
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "base,ethereum",
              "examples": ["base", "base,ethereum"]
            },
            "description": "Comma-separated supported chain slugs."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "minimum": 1,
              "maximum": 365
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "external", "internal", "erc20", "erc721", "erc1155"],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enriched transaction history after successful x402 payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxHistorySnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Payment required",
            "headers": {
              "payment-required": {
                "description": "Base64-encoded x402 payment requirements.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Server or provider error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PortfolioSnapshot": {
        "type": "object",
        "properties": {
          "address": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "chains": {
            "type": "array",
            "items": { "type": "string" }
          },
          "summary": {
            "$ref": "#/components/schemas/PortfolioSummary"
          },
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioToken"
            }
          },
          "positions": {
            "type": "array",
            "items": {}
          },
          "recentActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentActivity"
            }
          },
          "provider": {
            "type": "string",
            "const": "alchemy"
          }
        },
        "required": ["address", "timestamp", "chains", "summary", "tokens", "positions", "recentActivity", "provider"]
      },
...
      "TxHistorySnapshot": {
        "type": "object",
        "properties": {
          "address": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "chains": {
            "type": "array",
            "items": { "type": "string" }
          },
          "request": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TxHistoryTransaction"
            }
          },
          "pagination": {
            "type": "object"
          },
          "provider": {
            "type": "string",
            "const": "alchemy"
          },
          "note": { "type": "string" }
        },
        "required": ["address", "timestamp", "chains", "request", "summary", "transactions", "pagination", "provider", "note"]
      },
...
    }
  }
}